Update and rename TraceSquareCheck to TraceTimeCheck
This Code Checks if the matrix is square. If it is not, then an error message will be printed. If the matrix input is square, then it calculates the trace in 6 different ways AND the time each method took is recorded. Then a nice bar plot is created to display the data nicely.